Move services/{sync,indexer} to block-sync-renderer module
They're not used by other parts.
