+++ packages/neuron-wallet/src/block-sync-renderer/indexer/queue.ts
@@ -1,11 +1,11 @@
 import {  AddressPrefix } from '@nervosnetwork/ckb-sdk-utils'
 import ArrayUtils from 'utils/array'
 import logger from 'utils/logger'
-import GetBlocks from 'services/sync/get-blocks'
+import GetBlocks from 'block-sync-renderer/sync/get-blocks'
 import NetworksService from 'services/networks'
 import { Transaction, TransactionWithStatus } from 'types/cell-types'
 import TypeConvert from 'types/type-convert'
-import BlockNumber from 'services/sync/block-number'
+import BlockNumber from 'block-sync-renderer/sync/block-number'
 import LockUtils from 'models/lock-utils'
 import TransactionPersistor from 'services/tx/transaction-persistor'
 import IndexerTransaction from 'services/tx/indexer-transaction'
