+++ packages/neuron-wallet/src/block-sync-renderer/task.ts
@@ -1,8 +1,8 @@
 import { ipcRenderer } from 'electron'
 import initConnection from 'database/chain/ormconfig'
-import IndexerRPC from 'services/indexer/indexer-rpc'
-import BlockListener from 'services/sync/block-listener'
-import IndexerQueue from 'services/indexer/queue'
+import IndexerRPC from 'block-sync-renderer/indexer/indexer-rpc'
+import BlockListener from 'block-sync-renderer/sync/block-listener'
+import IndexerQueue from 'block-sync-renderer/indexer/queue'
 import LockUtils from 'models/lock-utils'
 import DaoUtils from 'models/dao-utils'
 import { register as registerTxStatusListener, unregister as unregisterTxStatusListener } from 'listeners/renderer/tx-status'
