+++ packages/neuron-wallet/src/listeners/renderer/tx-status.ts
@@ -6,7 +6,7 @@ import LockUtils from 'models/lock-utils'
 import { FailedTransaction, TransactionPersistor } from 'services/tx'
 import { CONNECTION_NOT_FOUND_NAME } from 'database/chain/ormconfig'
 import TypeConvert from 'types/type-convert'
-import GetBlocks from 'services/sync/get-blocks'
+import GetBlocks from 'block-sync-renderer/sync/get-blocks'
 import NetworksService from 'services/networks'
 import { AddressPrefix } from 'models/keys/address'
 import NodeService from 'services/node'
