+++ packages/neuron-wallet/src/services/wallets.ts
@@ -15,7 +15,7 @@ import NodeService from './node'
 import FileService from './file'
 import { TransactionsService, TransactionPersistor, TransactionGenerator } from './tx'
 import AddressService from './addresses'
-import { deindexLockHashes } from './indexer/deindex'
+import { deindexLockHashes } from '../block-sync-renderer/indexer/deindex'
 import NetworksService from 'services/networks'
 import AddressesService from 'services/addresses'
 import { Cell, DepType } from 'types/cell-types'
